refactor: move user cards from user list
