+++ packages/app/src/containers/UserListPage/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import FileUpload from '../../components/FileUpload';
 import Spinner from '../../components/Spinner';
+import UserCards from '../../components/UserCards';
 import UserList from '../../components/UserList';
 import useUsers from '../../hooks/useUsers';
 
@@ -12,6 +13,7 @@ function UserListPage(): JSX.Element {
     return (
         <div className="w-full md:w-3/4 mx-auto mt-12 px-4">
             <UserList users={response?.data.users} />
+            <UserCards className="mt-24" users={response?.data.users} />
             <div className="mt-24 mb-4 mx-auto max-w-md">
                 <FileUpload />
             </div>
